+++ CHANGELOG.md
@@ -5,6 +5,7 @@ _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob
 - **[UPDATE]** Updated to Kotlin 1.4.0.
 - **[UPDATE]** Updated to Kotlin Coroutines 1.3.9.
 - **[FEATURE]** (`ktx-scene2d`) Added `image` builders for `NinePatch`, `TextureRegion`, `Texture` and `Drawable`.
+- **[FEATURE]** (`ktx-collections`) Added alias `GdxArrayMap` for gds ArrayMap and `set` operator function to use square brackets assignment.
 
 #### 1.9.11-b1